Hacky solution for Canvas constantly resizing
Added a new `SizedCanvas`. This Canvas only resizes when its size changes which is controlled by the programmer. This also means that when you want to make it fit a container, this has to be done manually.

This solution will not be put into Tokamak as it is not SwiftUI compatible and better solutions exist. However, I do not know how to implement them.